+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+
+## ## [1.0.5] - 2021-1-4
+- isShowCircle: 是否显示边上的圆
+- onCircleCallback: 边上圆的点击回调
+- gridConfig: 自定义背景 {
+                      strokeColor: '#ccc', // 边的颜色
+                      strokeWidth: 1, // 边的宽度
+                      isLineDash: false, // 是否虚线显示
+                    }
+- 新增实例setEdgeValues方法： 
+    dfInstance?.setEdgeValues({ edgeId: 'id_001-id_002', text: 99 }
+
 ## [1.0.4] - 2021-1-1
 - feat：新增demo
 - fix: 修改控制栏不显示图标问题